Fix compiler warnings, and while at it, convert tabs to spaces, add copyright comment line to beginning of files etc. small stuff.

+++ EntityComponents/EC_3DCanvas/EC_3DCanvas.cpp
@@ -168,7 +168,7 @@ void EC_3DCanvas::Update()
         update_internals_ = false;
     }
 
-    if (texture->getWidth() != buffer.width() || texture->getHeight() != buffer.height())
+    if ((int)texture->getWidth() != buffer.width() || (int)texture->getHeight() != buffer.height())
     {
         texture->freeInternalResources();
         texture->setWidth(buffer.width());
